working on report api docs
